requirements.txt
asgiref==3.7.2
celery==5.3.6
Django==4.2.11
django-cors-headers==4.3.1
django-environ==0.11.2
django-filter==23.5
djangorestframework==3.14.0
kombu==5.3.4
psycopg2-binary==2.9.9
redis==5.0.3
requests==2.31.0
djangorestframework-simplejwt==5.3.1
PyJWT==2.8.0
pytest==7.4.4
pytest-django==4.7.0
sqlparse==0.4.4
tzdata==2023.3
